pool-model, gate-pool - file structure and tags fixed (#75)
* changes

* changes

* changes

* changes

* changes

* changes
